--- packages/commonwealth/client/scripts/views/pages/CreateQuest/CreateQuestForm/index.ts
@@ -1,3 +1,0 @@
-import CreateQuestForm from './CreateQuestForm';
-
-export default CreateQuestForm;
