+++ packages/commonwealth/client/scripts/views/pages/CreateQuest/QuestForm/QuestForm.scss
@@ -1,6 +1,6 @@
 @import '../../../../styles/shared';
 
-.CreateQuestForm {
+.QuestForm {
   display: flex;
   flex-direction: column;
   gap: 24px;
