+++ packages/commonwealth/client/scripts/views/pages/CreateQuest/QuestForm/QuestForm.tsx
@@ -15,12 +15,12 @@ import { CWForm } from 'views/components/component_kit/new_designs/CWForm';
 import { CWTextInput } from 'views/components/component_kit/new_designs/CWTextInput';
 import { withTooltip } from 'views/components/component_kit/new_designs/CWTooltip';
 import { CWRadioButton } from 'views/components/component_kit/new_designs/cw_radio_button';
-import './CreateQuestForm.scss';
 import QuestActionSubForm, { QuestAction } from './QuestActionSubForm';
-import useCreateQuestForm from './useCreateQuestForm';
+import './QuestForm.scss';
+import useQuestForm from './useQuestForm';
 import { questFormValidationSchema } from './validation';
 
-const CreateQuestForm = () => {
+const QuestForm = () => {
   const {
     addSubForm,
     questActionSubForms,
@@ -35,7 +35,7 @@ const CreateQuestForm = () => {
     minStartDate,
     repetitionCycleRadio,
     formMethodsRef,
-  } = useCreateQuestForm();
+  } = useQuestForm();
 
   return (
     <CWForm
@@ -43,7 +43,7 @@ const CreateQuestForm = () => {
       validationSchema={questFormValidationSchema}
       onSubmit={handleSubmit}
       onErrors={validateSubForms}
-      className="CreateQuestForm"
+      className="QuestForm"
     >
       <div className="quest-period-section">
         <div className="repeatition-selector">
@@ -194,4 +194,4 @@ const CreateQuestForm = () => {
   );
 };
 
-export default CreateQuestForm;
+export default QuestForm;
