+++ packages/commonwealth/client/scripts/views/pages/CreateQuest/QuestForm/index.ts
@@ -1,0 +1,3 @@
+import CreateQuestForm from './QuestForm';
+
+export default CreateQuestForm;
